Example013_RecursionAlgorithm

+++ Example013_RecursionAlgorithm/Program.cs
@@ -5,9 +5,9 @@
 // table [0,0] teble [0,1] table [0,2] table [0,4]
 // table [1,0] table [1,1] table [1,2] table [1,4]
 table[1, 2] = "слово";
-for (double rows = 0; rows < 2; rows++)
+for (int rows = 0; rows < 2; rows++)
 {
-    for (double columns = 0; columns < 5; columns++)
+    for (int columns = 0; columns < 5; columns++)
     {
         Console.Write($"-{table[rows, columns]}-");
     }
@@ -18,11 +18,11 @@ for (double rows = 0; rows < 2; rows++)
 
 
 
-/* void PrintArray(double[,] imege)
+/* void PrintArray(int[,] imege)
 {
-    for (double i = 0; i < imege.GetLength(0); i++)
+    for (int i = 0; i < imege.GetLength(0); i++)
     {
-        for (double j = 0; j < imege.GetLength(1); j++)
+        for (int j = 0; j < imege.GetLength(1); j++)
         {
             Console.Write($"{imege[i, j]} ");
         }
@@ -30,22 +30,23 @@ for (double rows = 0; rows < 2; rows++)
     }
     Console.WriteLine();
 }
-void FillArray(double[,] imege)
+void FillArray(int[,] imege)
 {
-    for (double i = 0; i < imege.GetLength(0); i++)
+    for (int i = 0; i < imege.GetLength(0); i++)
     {
-        for (double j = 0; j < imege.GetLength(1); j++)
+        for (int j = 0; j < imege.GetLength(1); j++)
         {
             imege[i, j] = new Random().Next(1, 10);
         }
     }
 }
-double[,] imegeix = new double[3, 4];
+int[,] imegeix = new int[3, 4];
 PrintArray(imegeix);
 FillArray(imegeix);
 PrintArray(imegeix);
 */
-/* double[,] pic = new double[,]
+
+/* int[,] pic = new int[,]
 {
 {0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
 {0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
@@ -73,11 +74,11 @@ PrintArray(imegeix);
 };
 
 
-void PrintImage(double[,] imege)
+void PrintImage(int[,] imege)
 {
-    for (double i = 0; i < imege.GetLength(0); i++)
+    for (int i = 0; i < imege.GetLength(0); i++)
     {
-        for (double j = 0; j < imege.GetLength(1); j++)
+        for (int j = 0; j < imege.GetLength(1); j++)
         {
             //Console.Write($"{imege[i, j]} ");
             if (imege[i, j] == 0) Console.Write($" ");
@@ -88,7 +89,7 @@ void PrintImage(double[,] imege)
     Console.WriteLine();
 
 }
-void FillImage(double row, double col)
+void FillImage(int row, int col)
 {
     if (pic[row, col] == 0)
     {
@@ -111,14 +112,14 @@ PrintImage(pic);
 //         4*3!
 //           3*2!
 
-/* double Factorial(double n)
+/* int Factorial(int n)
 {
     //1!=1
     //0!=1
     if (n == 1) return 1;
     else return n * Factorial(n - 1);
 }
-for (double i = 1; i <= 40; i++)
+for (int i = 1; i <= 40; i++)
 {
     Console.WriteLine($"{i}! -> {Factorial(i)}");
 }
@@ -131,30 +132,30 @@ for (double i = 1; i <= 40; i++)
 // f(2) = 1;
 // f(n) = f(n-1) + f(n-2);
 
-/* double Fibonacci(double n)
+/* int Fibonacci(int n)
 {
     if(n==1||n ==2) return 1;
     else return Fibonacci(n-1)+Fibonacci(n-2);
 }
-for (double i = 1; i <= 50; i ++)
+for (int i = 1; i <= 50; i ++)
 {
     Console.WriteLine($" F({i}) = {Fibonacci(i)};");
 }
 */
 /* Console.Clear();
 Console.Write("Введите число: ");
-double Num = Convert.ToInt32(Console.ReadLine());
+int Num = Convert.ToInt32(Console.ReadLine());
 Console.WriteLine();
 
 
 void MetodFibonacci()
 {
-    double fibonacci0 = 0;
-    double fibonacci1 = 1;
-    double fibonacciSum = 0;
+    int fibonacci0 = 0;
+    int fibonacci1 = 1;
+    int fibonacciSum = 0;
     Console.WriteLine($"F({0}) = {fibonacci0};"); 
     Console.WriteLine($"F({1}) = {fibonacci1};");
-    for (double i = 2; i <= Num; i++)
+    for (int i = 2; i <= Num; i++)
     {
         fibonacciSum = fibonacci0 + fibonacci1;
         fibonacci0 = fibonacci1;
